+++ example/lib/main.dart
@@ -27,7 +27,7 @@ class HomePage extends StatelessWidget {
         padding: EdgeInsets.all(16),
         children: <Widget>[
           Card(
-            child: RowLayout.card(
+            child: RowLayout(
               children: <Widget>[
                 RowItem.text('English', 'Hello'),
                 RowItem.text('Español', 'Hola'),
@@ -39,7 +39,7 @@ class HomePage extends StatelessWidget {
           ),
           Separator.cardSpacer(),
           Card(
-            child: RowLayout.card(
+            child: RowLayout(
               children: <Widget>[
                 RowItem.text(
                   'English',
@@ -71,7 +71,7 @@ class HomePage extends StatelessWidget {
           ),
           Separator.cardSpacer(),
           Card(
-            child: RowLayout.card(
+            child: RowLayout(
               children: <Widget>[
                 RowItem.icon('Firefox', true),
                 RowItem.icon('Chrome', false),
@@ -91,7 +91,7 @@ class HomePage extends StatelessWidget {
           ),
           Separator.cardSpacer(),
           Card(
-            child: RowLayout.card(
+            child: RowLayout(
               children: <Widget>[
                 RowItem.clickable(
                   'Android',