+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.2.0] - May 9th, 2019
+
+* **BREAKING CHANGE:** In `RowLayout`, substituted `card` named constructor to `cardList`.
+* **BREAKING CHANGE:** Constructor of `Separator.smallSpacer` doesn't accept parameters now.
+* Added most of `Column` parameters to `RowLayout`.
+* Changed `RowLayout.body` padding from `18` to `16`.
+* Changed `cardSpacer` height from `8` to `6`.
+* Updated example project.
+
 ## [0.1.0] - May 8th, 2019
 
 * **BREAKING CHANGE:** Changed `Separator.spacer` parameters to just `space`.